+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.75.1
 * Fixed return more than one row error if real error raised on try read next row
 * Fixed checking errors for session must be deleted
 * Changed signature of filter func in balancers (replaced argument from `conn.Conn` type to `endpoint.Info`)